feat: Scene+Entities panel.
This commit adds a simple scene hierarchy panel.
