+++ .github/workflows/cmake-build-cpp.yml
@@ -33,10 +33,6 @@ jobs:
             - uses: actions/setup-python@v4
               with:
                   python-version: "3.x"
-            - name: Get dependencies for shaderc
-              shell: bash
-              run: |
-                  python third_party/shaderc/utils/git-sync-deps
             - name: Install Linux Dependencies
               shell: cmake -P {0}
               run: |
@@ -107,7 +103,7 @@ jobs:
               uses: edvn0/setup-vulkan-sdk@v1.3.0
               with:
                   vulkan-query-version: latest
-                  vulkan-components: Vulkan-Headers, Vulkan-Loader
+                  vulkan-components: Vulkan-Headers, Vulkan-Loader, SPIRV-Cross, Glslang, SPIRV-Tools
                   vulkan-use-cache: true
             - name: Configure
               shell: cmake -P {0}
