+++ app/AlabasterLayer.hpp
@@ -38,16 +38,6 @@ struct AlabasterLayer final : public Alabaster::Layer {
 	void destroy() override;
 	void on_event(Alabaster::Event& event) override;
 
-	void draw_entity_node(SceneSystem::Entity& entity);
-	void draw_components(SceneSystem::Entity& entity);
-	template <SceneSystem::Component::IsComponent T> void display_add_component_entry(const std::string& entry_name)
-	{
-		if (!selected_entity.has_component<T>() && ImGui::MenuItem(entry_name.c_str())) {
-			selected_entity.add_component<T>();
-			ImGui::CloseCurrentPopup();
-		}
-	}
-
 private:
 	void handle_drag_drop();
