+++ libs/Core/include/core/Logger.hpp
@@ -34,32 +34,32 @@ namespace Alabaster {
 
 	namespace Log {
 
-		template <typename... Args> inline static constexpr auto info(auto&& fmt, Args&&... args) -> void
+		template <typename... Args> static constexpr auto info(auto&& fmt, Args&&... args) -> void
 		{
 			::Alabaster::Logger::get_core_logger()->info(fmt::vformat(fmt, fmt::make_format_args(std::forward<Args>(args)...)));
 		}
 
-		template <typename... Args> inline static constexpr auto debug(auto&& fmt, Args&&... args) -> void
+		template <typename... Args> static constexpr auto debug(auto&& fmt, Args&&... args) -> void
 		{
 			::Alabaster::Logger::get_core_logger()->debug(fmt::vformat(fmt, fmt::make_format_args(std::forward<Args>(args)...)));
 		}
 
-		template <typename... Args> inline static constexpr auto warn(auto&& fmt, Args&&... args) -> void
+		template <typename... Args> static constexpr auto warn(auto&& fmt, Args&&... args) -> void
 		{
 			::Alabaster::Logger::get_core_logger()->warn(fmt::vformat(fmt, fmt::make_format_args(std::forward<Args>(args)...)));
 		}
 
-		template <typename... Args> inline static constexpr auto error(auto&& fmt, Args&&... args) -> void
+		template <typename... Args> static constexpr auto error(auto&& fmt, Args&&... args) -> void
 		{
 			::Alabaster::Logger::get_core_logger()->error(fmt::vformat(fmt, fmt::make_format_args(std::forward<Args>(args)...)));
 		}
 
-		template <typename... Args> inline static constexpr auto critical(auto&& fmt, Args&&... args) -> void
+		template <typename... Args> static constexpr auto critical(auto&& fmt, Args&&... args) -> void
 		{
 			::Alabaster::Logger::get_core_logger()->critical(fmt::vformat(fmt, fmt::make_format_args(std::forward<Args>(args)...)));
 		}
 
-		template <typename... Args> inline static constexpr auto trace(auto&& fmt, Args&&... args) -> void
+		template <typename... Args> static constexpr auto trace(auto&& fmt, Args&&... args) -> void
 		{
 			::Alabaster::Logger::get_core_logger()->trace(fmt::vformat(fmt, fmt::make_format_args(std::forward<Args>(args)...)));
 		}
